Debug Functionality
Debug now appears when editor is open across all scenes. Note here that I may need to add a toggle to force debug on/off when editor is open as some of the Debugs significantly slow down the rendering/people may want to see the real content when the editor is open.